Retyped Lost Code
Sorry for my inattentiveness, but I accidentally deleted some of the
code you guys added by "merging" the branch master and attempting to
"reverting" it.  I retyped all of the code because I don't know how to
restore the repository back to its original state.   Could you guys
check to see if the code is accurate?  Thanks, and sorry again for
somewhat messing the code up.