A complete mess
Compass now takes in a VOR.

The plan was to have an event listener track left or right input, which
wiould increase/decrease the vor obs and needle variables.

What I can't figure out is how to have the frame refresh with the
updated compass.

Suggestion is to have an event listener into simulation class that will
do all the Compass updating and refreshing the frame.