Added Morse Code Portion to Simulated Radio
I'm not too sure how to implement the Morse Code part.  I just added a
"SetStationID" string that returns a string of the morse code given
three different letters.  I think Dr. Auernheimer just wants it printed
to the console, but I'm not sure, so I just had it return the String
instead.