Prints VOR station ID in the beginning . . .
I added two lines of code in the Simulation.java file that prints the
stationID to the screen.  I'm not sure if it's supposed to be manually
set, or randomly generated (such as detecting the nearest station), so
it's just pre-set for now.  In real life, it's supposed to be heard over
your headphones, so it doesn't affect the functionality of the VOR in
any way.